% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{entropic_fused_gromov_wasserstein}
\alias{entropic_fused_gromov_wasserstein}
\title{POT Alias for Entropic FGW}
\usage{
entropic_fused_gromov_wasserstein(...)
}
\value{
A list with \code{plan}, \code{fgw_dist}, \code{iterations}, \code{error}.
}
\description{
POT Alias for Entropic FGW
}
