% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{entropic_fused_gromov_wasserstein2}
\alias{entropic_fused_gromov_wasserstein2}
\title{POT Alias for Entropic FGW Value}
\usage{
entropic_fused_gromov_wasserstein2(...)
}
\value{
Numeric scalar FGW value.
}
\description{
POT Alias for Entropic FGW Value
}
