% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugw_factorized.R
\name{fugw_factorized2}
\alias{fugw_factorized2}
\title{Factorized FUGW Objective Value}
\usage{
fugw_factorized2(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{\link[=fugw_factorized]{fugw_factorized()}}.}
}
\value{
Numeric scalar matrix-free FUGW objective.
}
\description{
Factorized FUGW Objective Value
}
