% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgw.R
\name{fugw_kl2}
\alias{fugw_kl2}
\title{Fused Unbalanced Gromov-Wasserstein Objective Value}
\usage{
fugw_kl2(...)
}
\value{
Numeric scalar FUGW objective value.
}
\description{
Fused Unbalanced Gromov-Wasserstein Objective Value
}
