% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{fused_gromov_wasserstein}
\alias{fused_gromov_wasserstein}
\title{POT Alias for Exact FGW}
\usage{
fused_gromov_wasserstein(...)
}
\value{
A list with \code{plan}, \code{fgw_dist}, \code{iterations}, \code{error}, \code{loss_trace}.
}
\description{
POT Alias for Exact FGW
}
