% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{fused_unbalanced_gromov_wasserstein2}
\alias{fused_unbalanced_gromov_wasserstein2}
\title{POT Alias for Fused Unbalanced GW Value}
\usage{
fused_unbalanced_gromov_wasserstein2(...)
}
\value{
Numeric scalar FUGW value.
}
\description{
POT Alias for Fused Unbalanced GW Value
}
