% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiset.R
\name{multialign_normalize_plan}
\alias{multialign_normalize_plan}
\title{Normalize a Coupling Matrix}
\usage{
multialign_normalize_plan(
  P,
  mode = c("none", "row", "col", "sum"),
  eps = 1e-15
)
}
\arguments{
\item{P}{Coupling/transport matrix.}

\item{mode}{Normalization mode.}

\item{eps}{Stabilizer for division.}
}
\value{
Normalized coupling matrix.
}
\description{
Normalize a Coupling Matrix
}
