% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_utils.R
\name{ot_linear_cost}
\alias{ot_linear_cost}
\title{Linear transport cost}
\usage{
ot_linear_cost(M, plan)
}
\arguments{
\item{M}{Cost matrix.}

\item{plan}{Coupling or \code{rfugw_result}.}
}
\value{
Scalar \verb{<M, plan>}.
}
\description{
Linear transport cost
}
