% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api_ports.R
\name{partial_fused_gromov_wasserstein2}
\alias{partial_fused_gromov_wasserstein2}
\title{Partial Fused Gromov-Wasserstein Objective Value}
\usage{
partial_fused_gromov_wasserstein2(...)
}
\value{
Partial FGW value.
}
\description{
Partial Fused Gromov-Wasserstein Objective Value
}
