% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_barycenter.R
\name{print.rfugw_barycenter_result}
\alias{print.rfugw_barycenter_result}
\title{Print a fixed-support Wasserstein barycenter result}
\usage{
\method{print}{rfugw_barycenter_result}(x, ...)
}
\arguments{
\item{x}{An \code{rfugw_barycenter_result}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fixed-support Wasserstein barycenter result
}
