% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.rfugw_result}
\alias{print.rfugw_result}
\title{Print an rfugw result}
\usage{
\method{print}{rfugw_result}(x, ...)
}
\arguments{
\item{x}{An \code{rfugw_result}.}

\item{...}{Ignored.}
}
\description{
Print an rfugw result
}
