% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.summary.rfugw_result}
\alias{print.summary.rfugw_result}
\title{Print an rfugw result summary}
\usage{
\method{print}{summary.rfugw_result}(x, ...)
}
\arguments{
\item{x}{A \code{summary.rfugw_result}.}

\item{...}{Ignored.}
}
\description{
Print an rfugw result summary
}
