% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{rfugw_status}
\alias{rfugw_status}
\title{Extract solver status}
\usage{
rfugw_status(x)
}
\arguments{
\item{x}{An \code{rfugw_result}.}
}
\value{
Character status string.
}
\description{
Extract solver status
}
