% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api_ports.R
\name{semirelaxed_gromov_wasserstein2}
\alias{semirelaxed_gromov_wasserstein2}
\title{Semi-Relaxed Gromov-Wasserstein Objective Value}
\usage{
semirelaxed_gromov_wasserstein2(...)
}
\value{
Semirelaxed GW value.
}
\description{
Semi-Relaxed Gromov-Wasserstein Objective Value
}
