% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{summary.rfugw_result}
\alias{summary.rfugw_result}
\title{Summarize an rfugw result}
\usage{
\method{summary}{rfugw_result}(object, ...)
}
\arguments{
\item{object}{An \code{rfugw_result}.}

\item{...}{Ignored.}
}
\description{
Summarize an rfugw result
}
