% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{entropic_gromov_wasserstein2}
\alias{entropic_gromov_wasserstein2}
\title{Entropic Gromov-Wasserstein Objective Value}
\usage{
entropic_gromov_wasserstein2(...)
}
\value{
Numeric scalar GW value.
}
\description{
Entropic Gromov-Wasserstein Objective Value
}
