% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api_ports.R
\name{entropic_partial_gromov_wasserstein2}
\alias{entropic_partial_gromov_wasserstein2}
\title{Entropic Partial Gromov-Wasserstein Objective Value}
\usage{
entropic_partial_gromov_wasserstein2(...)
}
\value{
Entropic partial GW value.
}
\description{
Entropic Partial Gromov-Wasserstein Objective Value
}
