% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{entropic_semirelaxed_fused_gromov_wasserstein2}
\alias{entropic_semirelaxed_fused_gromov_wasserstein2}
\title{Entropic Semirelaxed Fused Gromov-Wasserstein Objective Value}
\usage{
entropic_semirelaxed_fused_gromov_wasserstein2(...)
}
\value{
Numeric scalar semirelaxed FGW value.
}
\description{
Entropic Semirelaxed Fused Gromov-Wasserstein Objective Value
}
