% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{entropic_semirelaxed_gromov_wasserstein2}
\alias{entropic_semirelaxed_gromov_wasserstein2}
\title{Entropic Semirelaxed Gromov-Wasserstein Objective Value}
\usage{
entropic_semirelaxed_gromov_wasserstein2(...)
}
\value{
Numeric scalar semirelaxed GW value.
}
\description{
Entropic Semirelaxed Gromov-Wasserstein Objective Value
}
