% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgw.R
\name{fgw_entropic2}
\alias{fgw_entropic2}
\title{Entropic Fused Gromov-Wasserstein Objective Value}
\usage{
fgw_entropic2(...)
}
\value{
Numeric scalar FGW value.
}
\description{
Entropic Fused Gromov-Wasserstein Objective Value
}
