% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api_ports.R
\name{fused_gromov_barycenters}
\alias{fused_gromov_barycenters}
\title{Fused Gromov-Wasserstein Barycenters}
\usage{
fused_gromov_barycenters(...)
}
\description{
Convenience alias to \code{entropic_fused_gromov_barycenters}.
}
