% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{fused_gromov_wasserstein2}
\alias{fused_gromov_wasserstein2}
\title{POT Alias for Exact FGW Value}
\usage{
fused_gromov_wasserstein2(...)
}
\value{
Numeric scalar FGW value.
}
\description{
POT Alias for Exact FGW Value
}
