% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api.R
\name{gromov_wasserstein2}
\alias{gromov_wasserstein2}
\title{Exact Gromov-Wasserstein Objective Value}
\usage{
gromov_wasserstein2(...)
}
\value{
Numeric scalar GW value.
}
\description{
Exact Gromov-Wasserstein Objective Value
}
