% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api_ports.R
\name{partial_gromov_wasserstein2}
\alias{partial_gromov_wasserstein2}
\title{Partial Gromov-Wasserstein Objective Value}
\usage{
partial_gromov_wasserstein2(...)
}
\value{
Partial GW value.
}
\description{
Partial Gromov-Wasserstein Objective Value
}
