% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot_api_ports.R
\name{semirelaxed_fused_gromov_wasserstein2}
\alias{semirelaxed_fused_gromov_wasserstein2}
\title{Semi-Relaxed Fused Gromov-Wasserstein Objective Value}
\usage{
semirelaxed_fused_gromov_wasserstein2(...)
}
\value{
Semirelaxed FGW value.
}
\description{
Semi-Relaxed Fused Gromov-Wasserstein Objective Value
}
